feat(phpunit) Decorrelate test suite from SUT.
atoum creates a default relation between a test suite and a class, they
are correlated. PHPUnit doesn't do that. If a test suite extends
`PHPUnit\Framework\TestCase`, then it is assumed that this relation must
be canceled, if a test suite extends `atoum\phpunit\test`, then the
relation is kept. The former is a child of the latter, so it is easy to
implement.

The constant and function mock engines, and the code coverage scores
will not work.